+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ---
 
+#### 30.10.2018
+
+* [Library] Override background visibility only if playing
+
 #### 29.10.2018
 
 * [Library] Implemented background playback as an option